Reenable chat test (#203706)
I think this isn't an issue with the test- every time it fails, we see "The Web Worker Extension Host did not start in 60s". The chat tests are the first tests, and so they just time out due to being affected by the slow startup. That EH timeout issue looks old.
Fix #203429