Enable shell integration in hidden terminals
As I understand it, hidden terminals were always exempty from shell
integration because hidden terminals were thought of as private to an
extension in a way, and also since they probably weren't going to be
interacted with. Especially since the introduction of the environment
variable collection API I think we should change this decision, not
doing so will lead to weirdness for extensions where hidden ones (which
may be shown later) aren't fully featured and get inconsistent envs.

Fixes #199611
Part of #197187