Offset terminal when inline chat appears to ensure cursor is visible
This will overlay the terminal inline chat widget on top of the terminal the same
as previously where the cursor would be visible. If it would not be visible, the
terminal wrapper element is now shifted upwards to ensure it's visible. This comes
with a couple of downsides:

- The very top of the terminal cannot be accessed as overflow is clipped by the
  view.
- The chat may overlap content if there is any below the cursor, which is an
  unlikely edge case.

This isn't as good as a zone widget-like system in xterm.js could be, but that is
quite the undertaking and complicates the renderer code quite a bit. So this is
a good compromise.

Fixes #4720