Fix opening links in the terminal with column numbers
If you open a link from the terminal like `src/foo.ext:5.0-4` then the link passed to `TerminalSearchLinkOpener.open()` will be `src/foo.ext:5.0-4`. It will try to match it against detected links in the line, but it will fail because the detect link has text `src/foo.ext`. Then it will pass the full `src/foo.ext:5.0-4` to the quick picker which doesn't understand that format.

This fixes the matching against detected links so it strips potential line/column numbers before matching against the detected links' texts.

This only seems to be necessary when the link path doesn't exactly match a file (e.g. if that file is in a subdirectory).