Remove cmd/ctrl+a in quick pick (#211893)
* Remove cmd/ctrl+a in quick pick

I don't think this is used much because it's honestly a broken experience. When you do cmd+a, it selects all items in the quick pick _and_ all of your code in the editor.

Since not a single issue has been opened around this, I'm just going to remove this hard coded keybinding.

* import