Sort @workspace /explain above @terminal /explain (#213482)
Fix microsoft/vscode-copilot-release#1245

This is a bit of a hack but has some advantanges over adding an 'order' field to the chat participant or something like that. We can keep the order the same in other cases except for when the user is typing to filter the / list, then @terminal slash commands will be pushed down.