Don't clear chat disposables too early while progressive rendering (#215907)
This clear() is wrong because this loop is incrementally updating content, not rerendering everything. Deleting is also not quite right because we are technically leaking when things get rerendered. But that is only the case while in the middle of streaming content, and things will always be fixed up when the response is done loading.
Fix #212853