Add experimental recursive version of executeHoverProvider command (#221993)
* Add test recursive version of command

For #218240

Adds a `executeHoverProvider_recursive` version of the command. This avoids calling any exclusive providers

Not super happy with this as it duplicates code for a very specialized use. However it does work and is considerably simpler than some of the other approaches I explored in #218240

Instead of adding duplicate commands, we could instead use an argument or somehow sneak this information in on the uri. Changing the uri is a little scary tho as we have to make sure the modified uri doesn't leak back and confuse providers

* Add versions for go to commands too