remote: gracefully disconnect from remote EH (#223422)
* remote: gracefully disconnect from remote EH

Addresses https://github.com/microsoft/vscode/issues/211462#issuecomment-2245716115

This applies the same fix seen in #218972 to the extension host. It adds
a `disconnect` method that is called and awaited during the lifecycle
shutdown, which is implemented by th remote extension host connection.

The disconnection is awaited, but I think this is safe because it only
awaits `.drain()` and does so in the context of methods which were
already async at higher levels.

* fix tests

* Revert "cli: honor --connection-token-file in serve-web (#219041)"

This reverts commit 33b66bd83e19ac76865266d4bf454f2c66f5d9d7. That
commit was incorrectly merged in the branch.