fix: `merge.toggleActiveConflict` as shortcuts (#225320)
Fixes #225319

When Merge Editor's actions are assigned to keyboard like this:

```json
[
    {
        "key": "ctrl+g a",
        "command": "merge.toggleActiveConflictInput1",
        "when": "isMergeEditor"
    },
    {
        "key": "ctrl+g b",
        "command": "merge.toggleActiveConflictInput2",
        "when": "isMergeEditor"
    },
    {
        "key": "ctrl+g ctrl+a",
        "command": "merge.acceptAllInput1",
        "when": "isMergeEditor"
    },
    {
        "key": "ctrl+g ctrl+b",
        "command": "merge.acceptAllInput2",
        "when": "isMergeEditor"
    },
]
```

Any of such action changes the state of the merged editor so diff range
becomes locked, i.e. even if a cursor is moved to another diff range the
actions target the locked one. If focus is changed the merge editor
returns to a normal state. This behavior is only reproducible if all the
actions are performed with keyboard because mouse clicking and calling
the actions from the command pallette changes the focus and everything
works correctly.