Fix focusResult of Search Editor (#205914)
Add focusResultsOnSearch option to Search Editor

The option allows to set where the focus(cursor) should be moved
in a use case where the Search Editor has been opened from the search
view ("Open in editor" button). The focus can be moved to the results
or to the input of Search Editor.