Fix empty username in chat (#228142)
There's a race where the chat participant is registered but hasn't set the username data yet (this requires a request to github). So, changing to always fall back to the persisted username.

I think this isn't 100% complete and I'll follow up as debt. I think we need to rerender the list on ChatAgentService.onDidChangeAgents (thought we already did this but I guess not?) or I think this is a good place to start using observables.