Try another fix for chat codicon styling issue
Previous `@layer` fix caused a regression

The root problem I was trying to work around was that there's no way to clear rules that are set in the global list style sheet. You can override them with more specific css rules, but you can't remove them so that other global css rules (such as those for setting file icon colors) work properly

Instead of making the default styles apply to every list, it seems like they should only apply the lists the don't have their own styling. I've change it so that you now have to use `default-styled-list` if you want the default styles to apply

This seems to work but there are a ton of places where we use lists. Open to other ideas too