More tweaks to inline anchor rendering (#229126)
Better centers the file icon and also supports more file icon types

As part of this I think I need to tweak the default css rules for the file icons. The issue is that for image file icons, if they are rendered inline with `content: " "` they can end up collapsing down to zero height. I can override this in my css rules, but that causes other issues with non-image icons

Instead we can use an em quad to make the icon take up the right height by default. This size can still be overridden by css rules elsewhere in the workbench