testing: fix double scrollbars on long output (#230417)
* testing: fix double scrollbars on long output

Previously we limited the editor height to 10,000 pixels for performance, but this could show a double scrollbar on long output. In this change, we instead maintain the editor size to be the size of the output peek / view and move the editor as scroll in the list moves to take advantage of the editor's virtualization capabilities.

Fixes #228991

* hygenie