[vscode] Decorations for #file is much better and does not break
The current implementation of the range detection is broken, especially for the case where the ranges are compared using the strats.

If we insert a new line before the #file symbol, the range is not updated properly (note previously we were just updating the column)

After this change, we can freely insert new lines anywhere and the ranges get updated. This required a bit of work on doing the range calculation properly, but helps with the user experience.