remote: broaden URI identification on openExternal (#234380)
I want to protocol-activate copilot in a remote workspace using
openExternal, but I observed the URI was being treated as a path and
malformed. This change broadens the URI identification logic to accept
anything that looks like a protocol. I think this is okay since we
were already excluding file URIs from the rebase, so this really just
impacts any other estoeric URIs in addition to the protocol `urlProtocol`.

fyi @alexdima