testing: allow enabling and disabling CR per-profile (#234713)
Adds a dropdown to the continous run button that allows enabling and
disabling profiles used in autorun individually. Pretty simple but
involved some refactoring of the CR service.

![](https://memes.peet.io/img/24-11-b6823379-2eb8-4fb1-bbeb-deb15670d441.png)

We can also technically do this on a per-test basis now too, although
this is not exposed in UI at the moment.