Turn local echo off by default and downgrade to preview
This feature has been a pain point for many remote users for a long
time. Something critical to the feature is that it must maintain
consistency and heal "mistakes", but unfortunately this isn't true and
it can lead to buffer corruption and therefore could lead the user to
run a command they didn't mean to.

Additionally we don't disable the feature after it fails (#119103) or
respect the ECHO termios mode and can therefore echo password
characters to the user (#130821).

My plan is to keep it off as a preview for the foreseeable future as
we don't have plans to prioritize this work any time soon.

Part of #126209