Add isEdit to ChatResponseCodeblockUriPart (#244190)
We need to distinguish between codeblocks that have URIs and should be rendered as codeblock pills, vs ones that should be rendered as normal codeblocks but get the "apply in foo.js" handling due to having URIs. Previously this was just based on the mode, but in order to render chat codeblocks and also edits codeblocks, we need this tracking of whether a codeblockuri represents an edited file, which comes from the intent invocation.

When we clean up codeblock-based edits in favor of tools all the time, we can clean up this crazy code...